requirements.txt
pyrad==2.0
-e git://github.com/andrewyager/django-radius.git#egg=django-radius
django-filter==0.11.0
djangorestframework==3.3.1
Markdown==2.6.4
django-admin-bootstrapped==2.5.6
django-bootstrap3==6.2.2
django-cors-headers==1.1.0
djangorestframework-jwt==1.7.2
django-rest-auth
django-configurations
python-dateutil